refactor: split entity_format.go by formatter concern
Move the cover formatter, binary_sensor translation table, and
availability shape into their own files so each domain or shared
concern has a clear home. Pure refactor — no behavior change, all
existing tests continue to pass against the same public surface.

Two motivations: \`entity_format.go\` was approaching the 500-line
architecture-guard threshold, and the upcoming domain formatters
(lock, media_player, fan, vacuum, update, climate enrichment) would
have pushed it well past. The per-domain file pattern scales
cleanly — each new formatter lives in its own \`entity_format_*.go\`
without the dispatcher growing.

Files:
  entity_format.go              — dispatcher + default/weather/climate/light/person/sun
  entity_format_availability.go — sentinel detection + unavailable + fetch_error
  entity_format_cover.go        — cover formatter
  entity_format_binary_sensor.go — device_class translation table + semanticState
